pyproject.toml
[project]
name = "thegaffer"
version = "0.1.0"
description = "A soccer tactical chat MCP agent that provides tactical advice and analysis"
authors = [
    {name = "TheGaffer Team", email = "tactics@thegaffer.com"}
]
dependencies = [
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "asyncio-mqtt>=0.16.0",
    "aiofiles>=23.0.0",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
packages = [{include = "thegaffer"}]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "pre-commit>=3.4.0",
]

[project.scripts]
thegaffer = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]
